requirements.txt
openai
replicate
litellm
python-dotenv
Pillow
moviepy
pydub
requests
numpy
